+++ bigml/model.py
@@ -53,18 +53,17 @@ LOGGER = logging.getLogger('BigML')
 
 import sys
 import locale
-import os
-import json
 
 from bigml.api import FINISHED
-from bigml.api import (get_status, error_message, BigML, get_model_id,
-                       check_resource)
-from bigml.util import (invert_dictionary, slugify, markdown_cleanup,
+from bigml.api import (error_message, BigML, get_model_id, get_status)
+from bigml.util import (slugify, markdown_cleanup,
                         prefix_as_comment, utf8,
                         find_locale, cast)
 from bigml.util import DEFAULT_LOCALE
 from bigml.tree import Tree
 from bigml.predicate import Predicate
+from bigml.basemodel import BaseModel, retrieve_model
+from bigml.basemodel import ONLY_MODEL, EXCLUDE_ROOT
 
 
 PYTHON_CONV = {
@@ -93,26 +92,6 @@ INDENT = u'    '
 
 STORAGE = './storage'
 
-ONLY_MODEL = 'only_model=true'
-EXCLUDE_ROOT = 'exclude=root;shorten=true'
-
-def retrieve_model(api, model_id, query_string=''):
-    """ Retrieves model info either from a local repo or from the remote server
-
-    """
-    if api.storage is not None:
-        try:
-            with open("%s%s%s" % (api.storage, os.sep,
-                                  model_id.replace("/", "_"))) as model_file:
-                model = json.loads(model_file.read())
-            return model
-        except ValueError:
-            raise ValueError("The file %s contains no JSON")
-        except IOError:
-            pass
-    model = check_resource(model_id, api.get_model, query_string)
-    return model
-
 
 def extract_objective(objective_field):
     """Extract the objective field id from the model structure
@@ -135,7 +114,7 @@ def print_importance(field_importance, fields, out=sys.stdout):
         count += 1
 
 
-class Model(object):
+class Model(BaseModel):
     """ A lightweight wrapper around a Tree model.
 
     Uses a BigML remote model to build a local version that can be used
@@ -143,114 +122,38 @@ class Model(object):
 
     """
 
-    def __init__(self, model, api=None, shorten=False):
+    def __init__(self, model, api=None):
 
-        query_string = ONLY_MODEL
-        if shorten:
-            query_string += ';%s' % EXCLUDE_ROOT
-        if (isinstance(model, dict) and 'resource' in model and
+
+        if not (isinstance(model, dict) and 'resource' in model and
                 model['resource'] is not None):
-            self.resource_id = model['resource']
-        else:
             if api is None:
-                storage = None if shorten else STORAGE
-                api = BigML(storage=storage)
+                api = BigML(storage=STORAGE)
             self.resource_id = get_model_id(model)
             if self.resource_id is None:
                 raise Exception(error_message(model,
                                               resource_type='model',
                                               method='get'))
+            query_string = ONLY_MODEL
             model = retrieve_model(api, self.resource_id,
                                    query_string=query_string)
-
+        BaseModel.__init__(self, model, api=api)
         if ('object' in model and isinstance(model['object'], dict)):
             model = model['object']
 
         if ('model' in model and isinstance(model['model'], dict)):
             status = get_status(model)
             if ('code' in status and status['code'] == FINISHED):
-                if 'model_fields' in model['model']:
-                    fields = model['model']['model_fields']
-                    # pagination or exclusion might cause a field not to
-                    # be in available fields dict
-                    if not all(key in model['model']['fields']
-                               for key in fields.keys()):
-                        raise Exception("Some fields are missing"
-                                        " to generate a local model."
-                                        " Please, provide a model with"
-                                        " the complete list of fields.")
-                    for field in fields:
-                        field_info = model['model']['fields'][field]
-                        if not shorten:
-                            fields[field]['summary'] = field_info['summary']
-                        fields[field]['name'] = field_info['name']
-                else:
-                    fields = model['model']['fields']
-                objective_field = model['objective_fields']
-                self.objective_field = extract_objective(objective_field)
-                self.uniquify_varnames(fields)
-                self.inverted_fields = invert_dictionary(fields)
-                self.all_inverted_fields = invert_dictionary(model['model']
-                                                             ['fields'])
-                if not shorten:
-                    self.tree = Tree(
-                        model['model']['root'],
-                        fields,
-                        self.objective_field)
-                self.fields = fields
-                self.description = model['description']
-                self.field_importance = model['model'].get('importance',
-                                                           None)
-                if self.field_importance:
-                    self.field_importance = [element for element
-                                             in self.field_importance
-                                             if element[0] in fields]
-                self.locale = model.get('locale', DEFAULT_LOCALE)
-
+                self.tree = Tree(
+                    model['model']['root'],
+                    self.fields,
+                    self.objective_field)
             else:
                 raise Exception("The model isn't finished yet")
         else:
             raise Exception("Cannot create the Model instance. Could not"
                             " find the 'model' key in the resource:\n\n%s" %
                             model)
-
-    def uniquify_varnames(self, fields):
-        """Tests if the fields names are unique. If they aren't, a
-           transformation is applied to ensure unicity.
-
-        """
-        unique_names = set([fields[key]['name'] for key in fields])
-        if len(unique_names) < len(fields):
-            self.transform_repeated_names(fields)
-
-    def transform_repeated_names(self, fields):
-        """If a field name is repeated, it will be transformed adding its
-           column number. If that combination is also a field name, the
-           field id will be added.
-
-        """
-        # The objective field treated first to avoid changing it.
-        unique_names = [fields[self.objective_field]['name']]
-
-        field_ids = [field_id for field_id in fields
-                     if field_id != self.objective_field]
-        for field_id in field_ids:
-            new_name = fields[field_id]['name']
-            if new_name in unique_names:
-                new_name = "{0}{1}".format(fields[field_id]['name'],
-                                           fields[field_id]['column_number'])
-                if new_name in unique_names:
-                    new_name = "{0}_{1}".format(new_name, field_id)
-                fields[field_id]['name'] = new_name
-            unique_names.append(new_name)
-
-            
-
-    def resource(self):
-        """Returns the model resource ID
-
-        """
-        return self.resource_id
 
     def fields(self, out=sys.stdout):
         """Describes and return the fields for this model.